Homework 35 1. Added retrofit for internet communication 2. Added tmdb connection to download films info 3. Tmdb api key is stored in Secret object in data folder which is not synchronised with git (added to gitignore) 4. Films info language is set to system language 5. Removed hard-coded film db 6. Removed unused resources.

+++ app/src/main/java/film/search/filmsearch/view/rvviewholders/FilmViewHolder.kt
@@ -26,7 +26,7 @@ class FilmViewHolder(val binding: FilmItemBinding) : RecyclerView.ViewHolder(bin
         binding.root.setOnClickListener {
             clickListener.click(film, position, binding)
         }
-        binding.ratingView.setProgress(film.rating)
+        binding.ratingView.setProgress(film.rating.toFloat())
         adapterPosition
     }
 }
